+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Updated deploy Action to use more modern processes.
 - Updated `PaginatedList` to be type-aware, showing which class is included in the response. (Thanks [@HandcartCactus](https://github.com/HandcartCactus))
+- Updated Sphinx
 
 ## [3.3.0] - 2023-08-27