+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - LTI Resource Links (Thanks, [@jsmnhou](https://github.com/jsmnhou))
 
 ### General
+
 - Added support for Python 3.12 and 3.13
 - Dropped support for Python 3.7 and 3.8