+++ CHANGELOG.md
@@ -5,11 +5,18 @@
 ### New Endpoint Coverage
 
 - LTI Resource Links (Thanks, [@jsmnhou](https://github.com/jsmnhou))
+- Smart Search API [BETA] (Thanks, [@alportoricensis](https://github.com/alportoricensis))
+
+### General
+
+- Added support for Python 3.12 and 3.13
+- Dropped support for Python 3.7 and 3.8
 
 ### Backstage
 
 - Updated deploy Action to use more modern processes.
 - Updated `PaginatedList` to be type-aware, showing which class is included in the response. (Thanks [@HandcartCactus](https://github.com/HandcartCactus))
+- Updated Sphinx
 
 ## [3.3.0] - 2023-08-27