+++ CHANGELOG.md
@@ -2,9 +2,23 @@
 
 ## [Unreleased]
 
+### New Endpoint Coverage
+
+- LTI Resource Links (Thanks, [@jsmnhou](https://github.com/jsmnhou))
+- Smart Search API [BETA] (Thanks, [@alportoricensis](https://github.com/alportoricensis))
+- New Quizzes Accommodations
+
+### General
+
+- Added support for Python 3.12 and 3.13
+- Dropped support for Python 3.7 and 3.8
+
 ### Backstage
 
 - Updated deploy Action to use more modern processes.
+- Updated `PaginatedList` to be type-aware, showing which class is included in the response. (Thanks [@HandcartCactus](https://github.com/HandcartCactus))
+- Updated Sphinx
+- Reworked how `Requester` handles JSON-only POST requests (currently, only New Quizzes Accommodations and GraphQL)
 
 ## [3.3.0] - 2023-08-27